LPS-52679 Move Servlet infrastructure to common module
This new module relies on Configuration Admin and Dependency Manager.

Dependency Manager is used for the user to be able to pinpoint
extensions that must be present in order for the CXF Bus to be created.

Due to CXF extension mechanism, these must be present when the bus is
created and can't be safely removed. That is why we are marking them as
required dependencies.

Each time a configuration is created using Configuration Admin a new CXF
Servlet is spawned in the given path using the given extensions. In case
any extension is changed the Servlet will be respawned with the new
extensions.